+++ packages/core/src/components/DatePicker/DatePicker.stories.tsx
@@ -1,6 +1,6 @@
 import { boolean, select, text } from '@storybook/addon-knobs';
 import React, { useState } from 'react';
-import { placement } from '../Popover/Popover.stories';
+import { placements } from '../Popover/Popover.stories';
 import { DatePicker } from './DatePicker';
 import { Props } from './types';
 
@@ -22,7 +22,7 @@ export const Basic = () => {
             label={text('Label', 'Date of Birth')}
             variant={select('variant', variants, 'outlined')}
             labelPosition={select('Label Position', labelPosition, 'left')}
-            popoverPlacement={select('Popover Placement', placement, 'bottom-start')}
+            popoverPlacement={select('Popover Placement', placements, 'bottom-start')}
             required={boolean('Required', false)}
             minSelectableDate={new Date(2020, 11, 1)}
             maxSelectableDate={new Date(2021, 11, 30)}