+++ packages/core/src/components/Popover/Popover.stories.tsx
@@ -1,11 +1,7 @@
 import { centerAligned, styled } from '@medly-components/utils';
-import { boolean, select } from '@storybook/addon-knobs';
-import React, { useCallback, useState } from 'react';
-import Popover from './Popover';
-import PopoverWrapper from './PopoverWrapper';
-import { Placement } from './PopoverWrapper/types';
+import { InteractionType, Placement } from './types';
 
-export const placement: Placement[] = [
+export const placements: Placement[] = [
     'top-start',
     'top',
     'top-end',
@@ -13,14 +9,16 @@ export const placement: Placement[] = [
     'right',
     'right-end',
     'bottom-start',
-    'bottom',   
+    'bottom',
     'bottom-end',
     'left-end',
     'left',
     'left-start'
 ];
 
-const DummyDiv = styled('div')`
+export const interactionTypes: InteractionType[] = ['hover', 'click'];
+
+export const DummyDiv = styled('div')`
     color: white;
     width: 200px;
     height: 100px;
@@ -30,57 +28,15 @@ const DummyDiv = styled('div')`
     ${centerAligned()};
 `;
 
-const DummyWrapper = styled('div')`
+export const DummyWrapper = styled('div')`
     width: 100%;
     height: 100%;
     ${centerAligned()};
 `;
 
-const DummyPopover = styled('div')`
+export const DummyPopover = styled('div')`
     color: white;
     padding: 1.5rem 2rem;
     background-color: grey;
     ${centerAligned()};
 `;
-
-export const WithClickInteraction = () => {
-    const [isPopoverVisible, setPopoverVisibilityState] = useState(false),
-        changePopoverState = useCallback(() => setPopoverVisibilityState(val => !val), []);
-
-    return (
-        <DummyWrapper>
-            <PopoverWrapper
-                showPopover={isPopoverVisible}
-                onClick={changePopoverState}
-                onOuterClick={changePopoverState}
-                placement={select('Placement', placement, 'top')}
-            >
-                <DummyDiv>Hello</DummyDiv>
-                <Popover fullWidth={boolean('Popover Full Width', false)} fullHeight={boolean('Popover Full Height', false)}>
-                    <DummyPopover>This is Popover</DummyPopover>
-                </Popover>
-            </PopoverWrapper>
-        </DummyWrapper>
-    );
-};
-
-export const WithHoverInteraction = () => {
-    const [isPopoverVisible, setPopoverVisibilityState] = useState(false),
-        changePopoverState = useCallback(() => setPopoverVisibilityState(val => !val), []);
-
-    return (
-        <DummyWrapper>
-            <PopoverWrapper
-                showPopover={isPopoverVisible}
-                onMouseEnter={changePopoverState}
-                onMouseLeave={changePopoverState}
-                placement={select('Placement', placement, 'top')}
-            >
-                <DummyDiv>Hello</DummyDiv>
-                <Popover fullWidth={boolean('Popover Full Width', false)} fullHeight={boolean('Popover Full Height', false)}>
-                    <DummyPopover>This is Popover</DummyPopover>
-                </Popover>
-            </PopoverWrapper>
-        </DummyWrapper>
-    );
-};