+++ packages/core/src/components/Popover/Popover.styled.tsx
@@ -1,0 +1,7 @@
+import { styled } from '@medly-components/utils';
+import { Props } from './types';
+
+export const Wrapper = styled('div')<Props>`
+    position: relative;
+    display: inline-flex;
+`;