+++ packages/core/src/components/Popover/Popover/types.ts
@@ -1,8 +1,0 @@
-import { HTMLProps } from '@medly-components/utils';
-
-export interface Props extends HTMLProps<HTMLDivElement> {
-    /** Takes full width of the parent component */
-    fullWidth?: boolean;
-    /** Takes full height of the parent component */
-    fullHeight?: boolean;
-}