+++ packages/core/src/components/Popover/PopoverWrapper/index.ts
@@ -1,3 +1,0 @@
-import { PopoverWrapper } from './PopoverWrapper';
-
-export default PopoverWrapper;