+++ packages/core/src/components/Popover/index.ts
@@ -1,3 +1,1 @@
-export { default as Popover } from './Popover';
-export { default as PopoverWithBackground } from './PopoverWithBackground';
-export { default as PopoverWrapper } from './PopoverWrapper';
+export { Popover as default } from './Popover';